requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2-binary
python-dotenv
